requirements.txt
django==3.0.5
django-allauth==0.41.0
djangorestframework==3.11.0
Pillow==7.1.1
django-cors-headers==3.2.1